# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/categories": {
            "post": {
                "tags": [
                    "categories"
                ],
                "summary": "Cerate categorie",
                "operationId": "categoryCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "موش و گربه"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "categorie created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "کودکان"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "dateTime",
                                            "example": "2023-01-19T14:09:23.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "dateTime",
                                            "example": "2023-01-19T14:09:23.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "put": {
                "tags": [
                    "categories"
                ],
                "summary": "Cerate categorie",
                "operationId": "categoryUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "categorie id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "کودکان"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "categorie updated succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "دسته بندی با موفقیت آپدیت شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "categories"
                ],
                "summary": "Delete categorie",
                "operationId": "categorieDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "categorie id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "categorie updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "دسته بندی با موفقیت حذف شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/questions": {
            "post": {
                "tags": [
                    "questions"
                ],
                "summary": "Cerate question",
                "operationId": "questionCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question",
                                    "answer",
                                    "test_id"
                                ],
                                "properties": {
                                    "question": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "mysql چند انجین دارد ؟ "
                                    },
                                    "answer": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "۲ انجین / Innodb / myIsam"
                                    },
                                    "test_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "question created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "question": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "mysql چند انجین دارد ؟ "
                                        },
                                        "answer": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "۲ انجین / Innodb / myIsam"
                                        },
                                        "test_id": {
                                            "type": "integer",
                                            "format": "integer",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "dateTime",
                                            "example": "2023-01-19T14:09:23.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "dateTime",
                                            "example": "2023-01-19T14:09:23.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/questions/{id}": {
            "put": {
                "tags": [
                    "questions"
                ],
                "summary": "Cerate question",
                "operationId": "questionUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "question id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question",
                                    "answer",
                                    "test_id"
                                ],
                                "properties": {
                                    "question": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "mysql چند انجین دارد ؟ "
                                    },
                                    "answer": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "۲ انجین / Innodb / myIsam"
                                    },
                                    "test_id": {
                                        "type": "integer",
                                        "format": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "question updated succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "درس با موفقیت آپدیت شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "questions"
                ],
                "summary": "Delete question",
                "operationId": "questionDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "question id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "question updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "درس با موفقیت حذف شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}